+++ tests/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-
 from django_peeringdb.models import LG_URLField, MultipleChoiceField, URLField